whisper-1/requirements.txt
fastapi
uvicorn[standard]
openai-whisper
piper
pathvalidate
pathlib
numpy
